添加图片裁剪上传

+++ auto-common/src/main/java/com/jw/base/BaseMapper.java
@@ -14,5 +14,5 @@ public interface BaseMapper<T> {
 	
 	T get(T t) throws Exception;
 	
-	Integer count(T t) throws Exception;
+	Long count(T t) throws Exception;
 }
